+++ sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/rx/ParallelDocumentQueryTest.java
@@ -485,10 +485,10 @@ public class ParallelDocumentQueryTest extends TestSuiteBase {
 
     private List<InternalObjectNode> prepareCosmosContainer(CosmosAsyncContainer cosmosContainer) {
         try {
-            truncateCollection(cosmosContainer);
+            cleanUpContainer(cosmosContainer);
         } catch (Throwable firstChanceException) {
             try {
-                truncateCollection(cosmosContainer);
+                cleanUpContainer(cosmosContainer);
             } catch (Throwable lastChanceException) {
                 String message = String.format("container %s truncation failed due to first chance %s followed by last chance %s",
                     cosmosContainer,
@@ -509,7 +509,7 @@ public class ParallelDocumentQueryTest extends TestSuiteBase {
             docDefList.add(getDocumentDefinition(99));
         }
 
-        List<InternalObjectNode> items = bulkInsertBlocking(cosmosContainer, docDefList);
+        List<InternalObjectNode> items = insertAllItemsBlocking(cosmosContainer, docDefList, true);
         waitIfNeededForReplicasToCatchUp(getClientBuilder());
         return items;
     }