+++ sdk/eventgrid/azure-messaging-eventgrid-systemevents/src/main/java/com/azure/messaging/eventgrid/systemevents/models/StorageBlobCreatedEventData.java
@@ -258,8 +258,13 @@ public final class StorageBlobCreatedEventData implements JsonSerializable<Stora
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
         jsonWriter.writeStringField("accessTier", this.accessTier == null ? null : this.accessTier.toString());
-        jsonWriter.writeMapField("storageDiagnostics", this.storageDiagnostics,
-            (writer, element) -> writer.writeUntyped(element == null ? null : element.toObject(Object.class)));
+        jsonWriter.writeMapField("storageDiagnostics", this.storageDiagnostics, (writer, element) -> {
+            if (element == null) {
+                writer.writeNull();
+            } else {
+                element.writeTo(writer);
+            }
+        });
         jsonWriter.writeStringField("api", this.api);
         jsonWriter.writeStringField("clientRequestId", this.clientRequestId);
         jsonWriter.writeStringField("requestId", this.requestId);